Add new client module
One client to rule them all! The new client contains all the
functionality necessary for communicating with drooms server.
There are three modes in which the client can run:
1) interactive (default) - used for user's interaction (registration,
new strategies/playgrounds/tournaments, tournament enrollment). This
part was taken from the old cli module and redisigned a bit. There is a
lot of work to be done.
2) execution - used for running the games on the client machines. Using
GameService interface it gets game queue, executes one of the games and
delivers the results back to server.
3) batch mode - will allow running the same commands as an interactive
client from the command line (e.g. adding playgrounds). This mode is not
yet implemented.